make code less disgusting
