+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "hyprsome"
 description = "A small CLI apps that allows to make Hyprland's workspaces work like Awesome in multi-monitor setup"
 license = "GPL-3.0"
-version = "0.1.3"
+version = "0.1.4"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
